:sparkles: feat: Add Layout with _app
