+++ apps/web/pages/_app.tsx
@@ -1,16 +1,23 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import { ApolloProvider, MuiProvider, SupabaseProvider } from "@misterdonut/ui";
+import {
+  ApolloProvider,
+  MuiProvider,
+  SupabaseProvider,
+  Layout,
+} from "@misterdonut/ui";
 
 const MyApp = ({ Component, pageProps }: AppProps) => (
-  <SupabaseProvider>
-    <ApolloProvider>
-      <MuiProvider>
-        <Component {...pageProps} />
-      </MuiProvider>
-    </ApolloProvider>
-  </SupabaseProvider>
+  <Layout>
+    <SupabaseProvider>
+      <ApolloProvider>
+        <MuiProvider>
+          <Component {...pageProps} />
+        </MuiProvider>
+      </ApolloProvider>
+    </SupabaseProvider>
+  </Layout>
 );
 
 export default MyApp;
