--- packages/ui/src/components/Layout/Chart/Chart.tsx
@@ -1,3 +1,0 @@
-import { Title } from "../Title";
-
-export const Chart = () => <Title>Today</Title>;
