--- packages/ui/src/components/Layout/Chart/index.ts
@@ -1,1 +1,0 @@
-export * from "./Chart";
