--- packages/ui/src/components/Layout/Dashboard/index.ts
@@ -1,1 +1,0 @@
-export * from "./Dashboard";
