--- packages/ui/src/components/Layout/Deposits/index.ts
@@ -1,1 +1,0 @@
-export * from "./Deposits";
