--- packages/ui/src/components/Layout/ListItems/index.ts
@@ -1,1 +1,0 @@
-export * from "./ListItems";
