--- packages/ui/src/components/Layout/Orders/Orders.test.tsx
@@ -1,9 +1,0 @@
-import { render } from "@testing-library/react";
-import { composeStories } from "@storybook/testing-react";
-import * as stories from "./Orders.stories";
-
-const { Default } = composeStories(stories);
-test("renders", () => {
-  render(<Default />);
-  expect(1).toBe(1);
-});
