--- packages/ui/src/components/Layout/Orders/index.ts
@@ -1,1 +1,0 @@
-export * from "./Orders";
