--- packages/ui/src/components/Layout/Title/index.ts
@@ -1,1 +1,0 @@
-export * from "./Title";
