+++ packages/ui/src/components/Layout/index.ts
@@ -1,7 +1,1 @@
-export * from "./Dashboard";
-export * from "./Chart";
-export * from "./Deposits";
-export * from "./Orders";
-export * from "./Title";
-export * from "./ListItems";
-// COMPONENT EXPORT
+export * from "./Layout";
