+++ packages/ui/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./Button";
+export * from "./Layout";
 // COMPONENT EXPORT
